Need to read config values from somewhere
Read config values from the config watcher

WIP

+++ core/services/ocr2/plugins/ocr2keeper/evmregistry/v21/logprovider/buffer.go
@@ -18,9 +18,9 @@ var (
 	// defaultFastExecLogsHigh is the default upper bound / maximum number of logs that Automation is committed to process for each upkeep,
 	// based on available capacity, i.e. if there are no logs from other upkeeps.
 	// Used by Log buffer to limit the number of logs we are saving in memory for each upkeep in a block
-	defaultFastExecLogsHigh = 32
+	defaultFastExecLogsHigh = int64(32)
 	// defaultNumOfLogUpkeeps is the default number of log upkeeps supported by the registry.
-	defaultNumOfLogUpkeeps = 50
+	defaultNumOfLogUpkeeps = int64(50)
 )
 
 // fetchedLog holds the log and the ID of the upkeep
